Switch Bootstrap to Tailwind CSS and fix bugs
Framework switch:
- Replace Bootstrap 4.3.1 CDN with Tailwind CSS CDN + custom config
- Add Inter font + Material Symbols Outlined icons
- Rewrite base.html with Tailwind design tokens from Stitch designs
- Rewrite header: glass navbar with blur, responsive mobile menu
- Rewrite footer: tonal surface design, no hard borders
- Add dark mode class support, drop-cap and scrollbar-hide styles

Bug fixes:
- Fix blog API URL ordering: move create/list above slug patterns
- Move ContactForm from blog/forms.py to personal/forms.py
- Remove unused DeleteBlogPost form and its import
- Replace hardcoded nyasablog.com URL with form GET action
