+++ blog/forms.py
@@ -1,6 +1,6 @@
 from django import forms
 
-from blog.models import BlogPost 
+from blog.models import BlogPost
 
 
 class CreateBlogPostForm(forms.ModelForm):
@@ -27,24 +27,3 @@ class UpdateBlogPostForm(forms.ModelForm):
 		if commit:
 			blog_post.save()
 		return blog_post
-
-class DeleteBlogPost(forms.ModelForm):
-	class Meta:
-		model = BlogPost
-		fields = ['title', 'body', 'image']
-
-def should_be_empty(value):
-    if value:
-        raise forms.ValidationError('Field is not empty')
-
-
-class ContactForm(forms.Form):
-    name = forms.CharField(max_length=80)
-    title = forms.CharField(max_length=80)
-    message = forms.CharField(widget=forms.Textarea)
-    email = forms.EmailField()
-    forcefield = forms.CharField(
-        required=False, widget=forms.HiddenInput, label="Leave empty", validators=[should_be_empty])
-
-	
-
