+++ templates/base.html
@@ -1,41 +1,99 @@
 {% load static %}
-
 <!DOCTYPE html>
-<html lang=en>
-
+<html class="light" lang="en">
 <head>
-	<!-- Required meta tags -->
- <title> Nyasa Blog</title>
- <link rel="shortcut icon" type="image/png" href="{% static 'nyasa.png' %}" %}">
-    <meta name="description" content="Nyasa Blog is a web application for Malawian influencers and content creators.
- ">
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-	<link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css" integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
-	
-</head>
-
-<body>
+<meta charset="utf-8"/>
+<meta name="viewport" content="width=device-width, initial-scale=1.0"/>
+<title>{% block title %}NyasaBlog | Editorial Excellence from Malawi{% endblock %}</title>
+<meta name="description" content="{% block meta_description %}NyasaBlog is the most trusted source for Malawian content — Culture, Entertainment, Entrepreneurship, and more.{% endblock %}"/>
+{% block og_meta %}{% endblock %}
+<link rel="shortcut icon" type="image/png" href="{% static 'nyasa.png' %}"/>
+<script src="https://cdn.tailwindcss.com?plugins=forms,container-queries"></script>
+<link href="https://fonts.googleapis.com/css2?family=Inter:wght@100..900&display=swap" rel="stylesheet"/>
+<link href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:wght,FILL@100..700,0..1&display=swap" rel="stylesheet"/>
+<script>
+tailwind.config = {
+  darkMode: "class",
+  theme: {
+    extend: {
+      colors: {
+        "background": "#f8f9fa",
+        "primary": "#002627",
+        "primary-container": "#0f3d3e",
+        "primary-fixed": "#beebeb",
+        "primary-fixed-dim": "#a3cfcf",
+        "on-primary": "#ffffff",
+        "on-primary-container": "#7da8a8",
+        "on-primary-fixed": "#002020",
+        "on-primary-fixed-variant": "#224d4e",
+        "secondary": "#4059aa",
+        "secondary-container": "#8fa7fe",
+        "secondary-fixed": "#dce1ff",
+        "secondary-fixed-dim": "#b6c4ff",
+        "on-secondary": "#ffffff",
+        "on-secondary-container": "#1d3989",
+        "on-secondary-fixed": "#00164e",
+        "on-secondary-fixed-variant": "#264191",
+        "tertiary": "#321d00",
+        "tertiary-container": "#4f3000",
+        "tertiary-fixed": "#ffddb8",
+        "tertiary-fixed-dim": "#ffb95f",
+        "on-tertiary": "#ffffff",
+        "on-tertiary-container": "#de8e00",
+        "on-tertiary-fixed": "#2a1700",
+        "on-tertiary-fixed-variant": "#653e00",
+        "surface": "#f8f9fa",
+        "surface-bright": "#f8f9fa",
+        "surface-dim": "#d9dadb",
+        "surface-tint": "#3b6566",
+        "surface-variant": "#e1e3e4",
+        "surface-container": "#edeeef",
+        "surface-container-low": "#f3f4f5",
+        "surface-container-high": "#e7e8e9",
+        "surface-container-highest": "#e1e3e4",
+        "surface-container-lowest": "#ffffff",
+        "on-surface": "#191c1d",
+        "on-surface-variant": "#404848",
+        "on-background": "#191c1d",
+        "outline": "#717978",
+        "outline-variant": "#c0c8c8",
+        "error": "#ba1a1a",
+        "error-container": "#ffdad6",
+        "on-error": "#ffffff",
+        "on-error-container": "#93000a",
+        "inverse-surface": "#2e3132",
+        "inverse-on-surface": "#f0f1f2",
+        "inverse-primary": "#a3cfcf"
+      },
+      fontFamily: {
+        "headline": ["Inter", "sans-serif"],
+        "body": ["Inter", "sans-serif"],
+        "label": ["Inter", "sans-serif"]
+      },
+      borderRadius: {
+        "DEFAULT": "0.25rem",
+        "lg": "0.5rem",
+        "xl": "0.75rem",
+        "full": "9999px"
+      },
+    },
+  },
+}
+</script>
 {% include 'snippets/base_css.html' %}
+{% block extra_head %}{% endblock %}
+</head>
+<body class="bg-background text-on-surface font-body selection:bg-secondary-container selection:text-on-secondary-container">
+
 {% include 'snippets/header.html' %}
 
-<!-- Body -->
-<style type="text/css">
-	.main{
-		min-height: 100vh;
-		height: 100%;
-	}
-</style>
-<div class="main">
-	{% block content %}
-
-	{% endblock content %}
-</div>
-<!-- End Body -->
+<main class="pt-24 pb-20 min-h-screen">
+{% block content %}
+{% endblock content %}
+</main>
 
 {% include 'snippets/footer.html' %}
 
+{% block extra_js %}{% endblock %}
 </body>
-
-
 </html>
