+++ templates/snippets/base_css.html
@@ -1,5 +1,19 @@
-<style type="text/css">
-	body{
-		background-color: #f2f2f2;
-	}
+<style>
+    .material-symbols-outlined {
+        font-variation-settings: 'FILL' 0, 'wght' 400, 'GRAD' 0, 'opsz' 24;
+    }
+    body { font-family: 'Inter', sans-serif; }
+    .drop-cap::first-letter {
+        float: left;
+        font-size: 5rem;
+        line-height: 1;
+        font-weight: 800;
+        padding-right: 0.75rem;
+        color: #002627;
+        font-family: 'Inter', sans-serif;
+    }
+    .hide-scrollbar::-webkit-scrollbar { display: none; }
+    .hide-scrollbar { -ms-overflow-style: none; scrollbar-width: none; }
+    .no-scrollbar::-webkit-scrollbar { display: none; }
+    .no-scrollbar { -ms-overflow-style: none; scrollbar-width: none; }
 </style>
